fix: nine defects clang-tidy found, three in the real-time path
Static analysis over this repository for the first time, with the same
config the core uses.

In the real-time path, all three widening bugs of the same shape --
a product computed in UInt32 and then used as a size_t byte count or
index:

- AES67IOHandler zeroed the output buffer with a byte count computed in
  UInt32. A guard above makes it unreachable today, and that guard is the
  only thing between a bad frame count and a wild memset inside the audio
  callback.
- Its two interleaving loops indexed with frame * channelCount in UInt32.
- RTPTransmitter sized its payload buffer with 3 * channels * frames in
  int.

Elsewhere:

- RTPReceiver and RTPTransmitter cleared their statistics with
  std::memset. Statistics is eleven std::atomic members, so that is
  undefined behaviour -- and those counters are read from another thread
  while the stream runs, which is exactly where it bites. Statistics
  already had a reset() doing it properly.
- Three destructors could let an exception out: ~StreamManager,
  ~SAPAnnouncer::Impl and ~MemoryGuard. All three run at teardown, where
  unwinding is plausible, and an exception escaping there is
  std::terminate.
- CoreAudioClockSource rounded elapsed nanoseconds with the (x + 0.5)
  cast the core had the same way.

Annotated rather than changed, each with its reason: three sizeof-on-
pointer in AudioClockDeviceList, which is what CoreAudio's API wants
since it writes a CFStringRef into the buffer, and the WITH_CLEANUP macro
argument that cannot be parenthesised because a lambda capture list will
not take a parenthesised name.

performance-avoid-endl is excluded: 59 findings, all in logging, and
endl flushes -- in a driver the last line before a crash is the evidence.

check-tidy.sh strips -arch from a copy of the compile database, because
this project builds universal and clang-tidy refuses a command with two
architectures. The real build stays universal.

Gate green: 10 suites, core contract, 38 files analysed with no
error-level findings.

+++ Driver/AES67IOHandler.cpp
@@ -114,7 +114,12 @@ void AES67IOHandler::processInput(float* outputData, UInt32 frameCount, UInt32 c
     float channelBuffer[kMaxFramesPerBuffer];
 
     if (frameCount > kMaxFramesPerBuffer) {
-        std::memset(outputData, 0, frameCount * channelCount * sizeof(float));
+        // size_t product: this is a byte count, and computing it in UInt32
+        // truncates before the widening. Guarded above, so unreachable today
+        // with sane values -- but this is the real-time path, and the guard is
+        // the only thing standing between a bad frame count and a wild memset.
+        std::memset(outputData, 0,
+                    static_cast<size_t>(frameCount) * static_cast<size_t>(channelCount) * sizeof(float));
         return;
     }
 
@@ -137,7 +142,7 @@ void AES67IOHandler::processInput(float* outputData, UInt32 frameCount, UInt32 c
         // Interleave into output
         // outputData layout: [ch0_f0, ch1_f0, ..., ch127_f0, ch0_f1, ch1_f1, ...]
         for (UInt32 frame = 0; frame < frameCount; ++frame) {
-            outputData[frame * channelCount + ch] = channelBuffer[frame];
+            outputData[static_cast<size_t>(frame) * channelCount + ch] = channelBuffer[frame];
         }
     }
 }
@@ -157,7 +162,7 @@ void AES67IOHandler::processOutput(const float* inputData, UInt32 frameCount, UI
 
     for (size_t ch = 0; ch < channelCount; ++ch) {
         for (UInt32 frame = 0; frame < frameCount; ++frame) {
-            channelBuffer[frame] = inputData[frame * channelCount + ch];
+            channelBuffer[frame] = inputData[static_cast<size_t>(frame) * channelCount + ch];
         }
 
         const size_t samplesWritten = outputBuffers[ch].write(channelBuffer, frameCount);
